Added types to support the `Tee` module concept.
Removed a ton of dead/vestigial functionality from LDAPRecords and renamed it to better reflect its intention. Added the list of userAccountControl values to facilitate reading them.

Added `readUserAccountControl` function.

Modified `Mold.getValue` to work the same as the `Filter` methods.

Realized I implemented `Filter.valuePresent` in a dumb way and made it simpler.

Added many new `Searcher` methods for things that are commonly wanted enough to be their own Searcher instead of the result of a Filter.

Finally, `Tee` will be a module to modify the behavior of `Filter`s such that you keep the original results list and do multiple different things with it.