Found byte arrays in the wild.
Not quite. What I actually discovered is that the type of items that come back from LDAP are...dynamic.

Eww.

So if something is a singleton string, it'll just be a string. If there are more strings, it will be a string array.

That makes "behave this way for this attribute" much harder.

Anyhow, byte arrays behave now.

I got rid of the panels from Spectre, Coff doesn't seem to expose the width control, and I don't want to mess with them manually (that's why I'm using Coff to begin with...) so I just ditched them. Got rid of a bodge as well.