Removed ADSIAttributes array, as I realized (entirely too late) that the `DirectoryServices.SearchResult` could be enumerated for the 'keys' it contains dynamically, rather than using an externally-provided list.
Oof.

Well, saves me the trouble of searching the contents of the RootDSE/schema container and dynamically re-creating the same results. This not only dramatically reduces the size of the codebase, but makes the program faster for not having to iterate over hundreds of entries that don't exist on any given SearchResult.

Updated the `PrettyPrinter` to crunchcase the match cases, as well as introducing a number-of-results line at the top of the output.